- Admin toimingute actionitesse lisatud kontroll, et kasutaja oleks ikka admin.

+++ controllers/FieldController.php
@@ -62,6 +62,10 @@ class FieldController extends Controller
      */
     public function actionCreate()
     {
+		if(!$this->IsAdmin())
+		{
+			return $this->render('error', ['name' => 'Not Found (#404)', 'message' => 'Puuduvad piisavad õigused.']);
+		}
         $model = new Field();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -81,6 +85,10 @@ class FieldController extends Controller
      */
     public function actionUpdate($id)
     {
+		if(!$this->IsAdmin())
+		{
+			return $this->render('error', ['name' => 'Not Found (#404)', 'message' => 'Puuduvad piisavad õigused.']);
+		}
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -100,6 +108,10 @@ class FieldController extends Controller
      */
     public function actionDelete($id)
     {
+		if(!$this->IsAdmin())
+		{
+			return $this->render('error', ['name' => 'Not Found (#404)', 'message' => 'Puuduvad piisavad õigused.']);
+		}
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
@@ -134,5 +146,16 @@ class FieldController extends Controller
 		$fields = Field::find()->where(['type_id' => $id])->all();
 		\Yii::$app->response->format = 'json';
 		return $fields;
+	}
+	
+	public function IsAdmin()
+	{
+		$identity = Yii::$app->user->identity;
+		$is_admin = false;
+		if(isset($identity))
+		{
+			$is_admin = $identity->isAdmin;	
+		}
+		return $is_admin;
 	}
 }
